🔨 Refactor: ProblemTable 디자인 수정 및 다른 곳에도 공통 컴포넌트로 적용
