+++ src/pages/Mypage/Mypage.vue
@@ -1,20 +1,7 @@
 <script setup>
 import Profile from "@/components/layout/Profile.vue";
-import { getCurrentGradeInfo } from "@/utils/getCurrentGradeInfo";
-import { ref, onMounted } from "vue";
 import MypageTabs from "./components/MypageTabs.vue";
 import Grade from "@/components/layout/Grade.vue";
-
-const points = ref(0);
-const gradeInfo = ref();
-
-onMounted(() => {
-  points.value = 0;
-  setTimeout(() => {
-    points.value = 40;
-    gradeInfo.value = getCurrentGradeInfo(40);
-  }, 100);
-});
 </script>
 <template>
   <div class="flex flex-col gap-16">
@@ -25,7 +12,7 @@ onMounted(() => {
         name="김세빈"
         email="rlatpqls13@gmail.com"
       />
-      <Grade :points="points" :gradeInfo="gradeInfo" />
+      <Grade />
     </section>
     <MypageTabs />
   </div>
